requirements.txt
azure-functions==1.12.0
instaloader==4.9.5
